requirements.txt
spacy
nltk
pandas
scikit-learn
azureml-sdk
apache-airflow
pyyaml
